Add rationale for the need for each gate.
